blackoilpolymerparams: introduce translation unit
move code for loading parameters from eclipse state into it
