+++ opm/models/blackoil/blackoilpolymermodules.hh
@@ -39,15 +39,6 @@
 
 #include <opm/models/utils/propertysystem.hh>
 
-#if HAVE_ECL_INPUT
-#include <opm/input/eclipse/EclipseState/EclipseState.hpp>
-#include <opm/input/eclipse/EclipseState/Tables/PlyadsTable.hpp>
-#include <opm/input/eclipse/EclipseState/Tables/PlymaxTable.hpp>
-#include <opm/input/eclipse/EclipseState/Tables/PlyrockTable.hpp>
-#include <opm/input/eclipse/EclipseState/Tables/PlyshlogTable.hpp>
-#include <opm/input/eclipse/EclipseState/Tables/PlyviscTable.hpp>
-#endif
-
 #include <cmath>
 #include <stdexcept>
 #include <string>
@@ -93,259 +84,11 @@ class BlackOilPolymerModule
     static constexpr unsigned numPhases = FluidSystem::numPhases;
 
 public:
-#if HAVE_ECL_INPUT
-    /*!
-     * \brief Initialize all internal data structures needed by the polymer module
-     */
-    static void initFromState(const EclipseState& eclState)
+    //! \brief Set parameters.
+    static void setParams(BlackOilPolymerParams<Scalar>&& params)
     {
-        // some sanity checks: if polymers are enabled, the POLYMER keyword must be
-        // present, if polymers are disabled the keyword must not be present.
-        if (enablePolymer && !eclState.runspec().phases().active(Phase::POLYMER)) {
-            throw std::runtime_error("Non-trivial polymer treatment requested at compile time, but "
-                                     "the deck does not contain the POLYMER keyword");
-        }
-        else if (!enablePolymer && eclState.runspec().phases().active(Phase::POLYMER)) {
-            throw std::runtime_error("Polymer treatment disabled at compile time, but the deck "
-                                     "contains the POLYMER keyword");
-        }
-
-        if (enablePolymerMolarWeight && !eclState.runspec().phases().active(Phase::POLYMW)) {
-            throw std::runtime_error("Polymer molecular weight tracking is enabled at compile time, but "
-                                     "the deck does not contain the POLYMW keyword");
-        }
-        else if (!enablePolymerMolarWeight && eclState.runspec().phases().active(Phase::POLYMW)) {
-            throw std::runtime_error("Polymer molecular weight tracking is disabled at compile time, but the deck "
-                                     "contains the POLYMW keyword");
-        }
-
-        if (enablePolymerMolarWeight && !enablePolymer) {
-            throw std::runtime_error("Polymer molecular weight tracking is enabled while polymer treatment "
-                                     "is disabled at compile time");
-        }
-
-        if (!eclState.runspec().phases().active(Phase::POLYMER))
-            return; // polymer treatment is supposed to be disabled
-
-        const auto& tableManager = eclState.getTableManager();
-
-        unsigned numSatRegions = tableManager.getTabdims().getNumSatTables();
-        params_.setNumSatRegions(numSatRegions);
-
-        // initialize the objects which deal with the PLYROCK keyword
-        const auto& plyrockTables = tableManager.getPlyrockTables();
-        if (!plyrockTables.empty()) {
-            assert(numSatRegions == plyrockTables.size());
-            for (unsigned satRegionIdx = 0; satRegionIdx < numSatRegions; ++ satRegionIdx) {
-                const auto& plyrockTable = plyrockTables.template getTable<PlyrockTable>(satRegionIdx);
-                params_.setPlyrock(satRegionIdx,
-                                   plyrockTable.getDeadPoreVolumeColumn()[0],
-                                   plyrockTable.getResidualResistanceFactorColumn()[0],
-                                   plyrockTable.getRockDensityFactorColumn()[0],
-                                   static_cast<typename BlackOilPolymerParams<Scalar>::AdsorptionBehaviour>(plyrockTable.getAdsorbtionIndexColumn()[0]),
-                                   plyrockTable.getMaxAdsorbtionColumn()[0]);
-            }
-        }
-        else {
-            throw std::runtime_error("PLYROCK must be specified in POLYMER runs\n");
-        }
-
-        // initialize the objects which deal with the PLYADS keyword
-        const auto& plyadsTables = tableManager.getPlyadsTables();
-        if (!plyadsTables.empty()) {
-            assert(numSatRegions == plyadsTables.size());
-            for (unsigned satRegionIdx = 0; satRegionIdx < numSatRegions; ++ satRegionIdx) {
-                const auto& plyadsTable = plyadsTables.template getTable<PlyadsTable>(satRegionIdx);
-                // Copy data
-                const auto& c = plyadsTable.getPolymerConcentrationColumn();
-                const auto& ads = plyadsTable.getAdsorbedPolymerColumn();
-                params_.plyadsAdsorbedPolymer_[satRegionIdx].setXYContainers(c, ads);
-            }
-        }
-        else {
-            throw std::runtime_error("PLYADS must be specified in POLYMER runs\n");
-        }
-
-
-        unsigned numPvtRegions = tableManager.getTabdims().getNumPVTTables();
-        params_.plyviscViscosityMultiplierTable_.resize(numPvtRegions);
-
-        // initialize the objects which deal with the PLYVISC keyword
-        const auto& plyviscTables = tableManager.getPlyviscTables();
-        if (!plyviscTables.empty()) {
-            // different viscosity model is used for POLYMW
-            if (enablePolymerMolarWeight) {
-                OpmLog::warning("PLYVISC should not be used in POLYMW runs, "
-                                     "it will have no effect. A viscosity model based on PLYVMH is used instead.\n");
-            }
-            else {
-                assert(numPvtRegions == plyviscTables.size());
-                for (unsigned pvtRegionIdx = 0; pvtRegionIdx < numPvtRegions; ++ pvtRegionIdx) {
-                    const auto& plyadsTable = plyviscTables.template getTable<PlyviscTable>(pvtRegionIdx);
-                    // Copy data
-                    const auto& c = plyadsTable.getPolymerConcentrationColumn();
-                    const auto& visc = plyadsTable.getViscosityMultiplierColumn();
-                    params_.plyviscViscosityMultiplierTable_[pvtRegionIdx].setXYContainers(c, visc);
-                }
-            }
-        }
-        else if (!enablePolymerMolarWeight) {
-            throw std::runtime_error("PLYVISC must be specified in POLYMER runs\n");
-        }
-
-        // initialize the objects which deal with the PLYMAX keyword
-        const auto& plymaxTables = tableManager.getPlymaxTables();
-        const unsigned numMixRegions = plymaxTables.size();
-        params_.setNumMixRegions(numMixRegions, enablePolymerMolarWeight);
-        if (!plymaxTables.empty()) {
-            for (unsigned mixRegionIdx = 0; mixRegionIdx < numMixRegions; ++ mixRegionIdx) {
-                const auto& plymaxTable = plymaxTables.template getTable<PlymaxTable>(mixRegionIdx);
-                params_.plymaxMaxConcentration_[mixRegionIdx] = plymaxTable.getPolymerConcentrationColumn()[0];
-            }
-        }
-        else {
-            throw std::runtime_error("PLYMAX must be specified in POLYMER runs\n");
-        }
-
-        if (!eclState.getTableManager().getPlmixparTable().empty()) {
-            if (enablePolymerMolarWeight) {
-                OpmLog::warning("PLMIXPAR should not be used in POLYMW runs, it will have no effect.\n");
-            }
-            else {
-                const auto& plmixparTable = eclState.getTableManager().getPlmixparTable();
-                // initialize the objects which deal with the PLMIXPAR keyword
-                for (unsigned mixRegionIdx = 0; mixRegionIdx < numMixRegions; ++ mixRegionIdx) {
-                    params_.plymixparToddLongstaff_[mixRegionIdx] = plmixparTable[mixRegionIdx].todd_langstaff;
-                }
-            }
-        }
-        else if (!enablePolymerMolarWeight) {
-            throw std::runtime_error("PLMIXPAR must be specified in POLYMER runs\n");
-        }
-
-        params_.hasPlyshlog_ = eclState.getTableManager().hasTables("PLYSHLOG");
-        params_.hasShrate_ = eclState.getTableManager().useShrate();
-
-        if ((params_.hasPlyshlog_ || params_.hasShrate_) && enablePolymerMolarWeight) {
-            OpmLog::warning("PLYSHLOG and SHRATE should not be used in POLYMW runs, they will have no effect.\n");
-        }
-
-        if (params_.hasPlyshlog_ && !enablePolymerMolarWeight) {
-            const auto& plyshlogTables = tableManager.getPlyshlogTables();
-            assert(numPvtRegions == plyshlogTables.size());
-            params_.plyshlogShearEffectRefMultiplier_.resize(numPvtRegions);
-            params_.plyshlogShearEffectRefLogVelocity_.resize(numPvtRegions);
-            for (unsigned pvtRegionIdx = 0; pvtRegionIdx < numPvtRegions; ++ pvtRegionIdx) {
-                const auto& plyshlogTable = plyshlogTables.template getTable<PlyshlogTable>(pvtRegionIdx);
-
-                Scalar plyshlogRefPolymerConcentration = plyshlogTable.getRefPolymerConcentration();
-                auto waterVelocity = plyshlogTable.getWaterVelocityColumn().vectorCopy();
-                auto shearMultiplier = plyshlogTable.getShearMultiplierColumn().vectorCopy();
-
-                // do the unit version here for the waterVelocity
-                UnitSystem unitSystem = eclState.getDeckUnitSystem();
-                double siFactor = params_.hasShrate_? unitSystem.parse("1/Time").getSIScaling() : unitSystem.parse("Length/Time").getSIScaling();
-                for (size_t i = 0; i < waterVelocity.size(); ++i) {
-                    waterVelocity[i] *= siFactor;
-                    // for plyshlog the input must be stored as logarithms
-                    // the interpolation is then done the log-space.
-                    waterVelocity[i] = std::log(waterVelocity[i]);
-                }
-
-                Scalar refViscMult = params_.plyviscViscosityMultiplierTable_[pvtRegionIdx].eval(plyshlogRefPolymerConcentration, /*extrapolate=*/true);
-                // convert the table using referece conditions
-                for (size_t i = 0; i < waterVelocity.size(); ++i) {
-                    shearMultiplier[i] *= refViscMult;
-                    shearMultiplier[i] -= 1;
-                    shearMultiplier[i] /= (refViscMult - 1);
-                    shearMultiplier[i] = shearMultiplier[i];
-                }
-                params_.plyshlogShearEffectRefMultiplier_[pvtRegionIdx].resize(waterVelocity.size());
-                params_.plyshlogShearEffectRefLogVelocity_[pvtRegionIdx].resize(waterVelocity.size());
-
-                for (size_t i = 0; i < waterVelocity.size(); ++i) {
-                    params_.plyshlogShearEffectRefMultiplier_[pvtRegionIdx][i] = shearMultiplier[i];
-                    params_.plyshlogShearEffectRefLogVelocity_[pvtRegionIdx][i] = waterVelocity[i];
-                }
-            }
-        }
-
-        if (params_.hasShrate_ && !enablePolymerMolarWeight) {
-            if (!params_.hasPlyshlog_) {
-                throw std::runtime_error("PLYSHLOG must be specified if SHRATE is used in POLYMER runs\n");
-            }
-            const auto& shrateTable = eclState.getTableManager().getShrateTable();
-            params_.shrate_.resize(numPvtRegions);
-            for (unsigned pvtRegionIdx = 0; pvtRegionIdx < numPvtRegions; ++ pvtRegionIdx) {
-                if (shrateTable.empty()) {
-                    params_.shrate_[pvtRegionIdx] = 4.8; //default;
-                }
-                else if (shrateTable.size() == numPvtRegions) {
-                    params_.shrate_[pvtRegionIdx] = shrateTable[pvtRegionIdx].rate;
-                }
-                else {
-                    throw std::runtime_error("SHRATE must either have 0 or number of NUMPVT entries\n");
-                }
-            }
-        }
-
-        if constexpr (enablePolymerMolarWeight) {
-            const auto& plyvmhTable = eclState.getTableManager().getPlyvmhTable();
-            if (!plyvmhTable.empty()) {
-                assert(plyvmhTable.size() == numMixRegions);
-                for (size_t regionIdx = 0; regionIdx < numMixRegions; ++regionIdx) {
-                    params_.plyvmhCoefficients_[regionIdx].k_mh = plyvmhTable[regionIdx].k_mh;
-                    params_.plyvmhCoefficients_[regionIdx].a_mh = plyvmhTable[regionIdx].a_mh;
-                    params_.plyvmhCoefficients_[regionIdx].gamma = plyvmhTable[regionIdx].gamma;
-                    params_.plyvmhCoefficients_[regionIdx].kappa = plyvmhTable[regionIdx].kappa;
-                }
-            }
-            else {
-                throw std::runtime_error("PLYVMH keyword must be specified in POLYMW rus \n");
-            }
-
-            // handling PLYMWINJ keyword
-            const auto& plymwinjTables = tableManager.getPlymwinjTables();
-            for (const auto& table : plymwinjTables) {
-                const int tableNumber = table.first;
-                const auto& plymwinjtable = table.second;
-                const std::vector<double>& throughput = plymwinjtable.getThroughputs();
-                const std::vector<double>& watervelocity = plymwinjtable.getVelocities();
-                const std::vector<std::vector<double>>& molecularweight = plymwinjtable.getMoleWeights();
-                TabulatedTwoDFunction tablefunc(throughput, watervelocity, molecularweight, true, false);
-                params_.plymwinjTables_[tableNumber] = std::move(tablefunc);
-            }
-
-            // handling SKPRWAT keyword
-            const auto& skprwatTables = tableManager.getSkprwatTables();
-            for (const auto& table : skprwatTables) {
-                const int tableNumber = table.first;
-                const auto& skprwattable = table.second;
-                const std::vector<double>& throughput = skprwattable.getThroughputs();
-                const std::vector<double>& watervelocity = skprwattable.getVelocities();
-                const std::vector<std::vector<double>>& skinpressure = skprwattable.getSkinPressures();
-                TabulatedTwoDFunction tablefunc(throughput, watervelocity, skinpressure, true, false);
-                params_.skprwatTables_[tableNumber] = std::move(tablefunc);
-            }
-
-            // handling SKPRPOLY keyword
-            const auto& skprpolyTables = tableManager.getSkprpolyTables();
-            for (const auto& table : skprpolyTables) {
-                const int tableNumber = table.first;
-                const auto& skprpolytable = table.second;
-                const std::vector<double>& throughput = skprpolytable.getThroughputs();
-                const std::vector<double>& watervelocity = skprpolytable.getVelocities();
-                const std::vector<std::vector<double>>& skinpressure = skprpolytable.getSkinPressures();
-                const double refPolymerConcentration = skprpolytable.referenceConcentration();
-                typename BlackOilPolymerParams<Scalar>::SkprpolyTable tablefunc =
-                    {refPolymerConcentration,
-                     TabulatedTwoDFunction(throughput, watervelocity, skinpressure, true, false)};
-                params_.skprpolyTables_[tableNumber] = std::move(tablefunc);
-            }
-        }
+        params_ = params;
     }
-#endif
-
     /*!
     * \brief get the PLYMWINJ table
     */
